pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "s5-shb-agent"
version = "0.1.0"
description = "Society 5.0 Agentic Blockchain Smart Home Agent"
requires-python = ">=3.11"

dependencies = [
    # Engine core
    "google-generativeai>=0.8.0",
    "cryptography>=42.0.0",
    "fastmcp>=2.0.0",
    "scikit-learn>=1.0.0",
    "numpy>=1.24.0",
    # Web layer
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
voice = [
    "SpeechRecognition>=3.10.0",
    "PyAudio>=0.2.14",
    "pyttsx3>=2.90",
]
dl = [
    "torch>=2.0.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]
eval = [
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "jupyter",
    "ipykernel",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["engine*", "web*"]
